requirements.txt
streamlit
load_dotenv
google-generativeai